Don't effectively blacklist pruned nodes.
If upon connecting it's found that the node doesn't have the blocks we need, we can always disconenct then, but unless we give the node a chance to become a full-node, it will forever be forgotten about. Also, even pruned nodes provide much use for block propagation and it's only during IBD that we're needing historical blocks anyway. Perhaps in this situation it is reasonable to skip nodes that were once seen to be pruning (but may no longer be).