Disallow fingerprinting from blocktxn and cmpctblocks.
Currently we do not send orphan blocks that are older than one month for
block requests, but this logic is was not present when responding to
getblocktxn request or getdata cmpctblock requests.

Ensure that we can also not be fingerprinted with these requests, by
using chainwork instead of height as the criteria.

Another way to close this attack vector is to not accept blocks received that
are are unrequested and have less nChainWork than our ActiveChain's Tip.